Breaking test utilities apart, segregating them from source code
