+++ packages/wallet/test/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './generateTestWallet';
+export * from './seedTestWallet';
